+++ tests/commands/nd_material/test_contact.py
@@ -7,6 +7,6 @@ def test_contact_material2d():
 
 
 def test_contact_material3d():
-    osi = o3.OpenSeesInstance(ndm=2)
+    osi = o3.OpenSeesInstance(ndm=3)
     o3.nd_material.ContactMaterial3D(osi, mu=1.0, g_mod=1.0, c=1.0, t=1.0)