+++ HISTORY.rst
@@ -12,6 +12,7 @@ History
 * Added truss element objects
 * Fixed issue with BeamOnNonlinearWinklerFoundation where fixities were not applied to base node.
 * Fixed issue with getting element node tags when there is only one element
+* Fixed issues with Contact elements